fix(calendar): remove unused code path
