+++ services/calendar/occurrences.go
@@ -2,256 +2,16 @@ package calendar
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"sort"
-	"strings"
 	"time"
 
 	calendarresource "github.com/fivenet-app/fivenet/v2026/gen/go/proto/resources/calendar"
 	calendarentries "github.com/fivenet-app/fivenet/v2026/gen/go/proto/resources/calendar/entries"
 	"github.com/fivenet-app/fivenet/v2026/gen/go/proto/resources/timestamp"
 	"github.com/fivenet-app/fivenet/v2026/gen/go/proto/resources/userinfo"
-	"github.com/fivenet-app/fivenet/v2026/query/fivenet/table"
-	"github.com/go-jet/jet/v2/mysql"
-	"github.com/go-jet/jet/v2/qrm"
 	"google.golang.org/protobuf/proto"
 )
-
-const (
-	birthdayCalendarColor = "neutral"
-)
-
-type birthdayColleague struct {
-	UserID      int32  `alias:"user_id"`
-	Firstname   string `alias:"firstname"`
-	Lastname    string `alias:"lastname"`
-	Dateofbirth string `alias:"dateofbirth"`
-}
-
-func (s *Server) ensureJobBirthdayCalendar(
-	ctx context.Context,
-	userInfo *userinfo.UserInfo,
-) (*calendarresource.Calendar, error) {
-	job := strings.TrimSpace(userInfo.GetJob())
-	if job == "" {
-		return nil, nil
-	}
-
-	jobInfo := s.enricher.GetJobByName(job)
-	title := birthdayCalendarTitle(s.i18n, s.appCfg, job, jobInfo)
-
-	existing, err := s.getCalendar(
-		ctx,
-		userInfo,
-		mysql.AND(
-			tCalendar.DeletedAt.IS_NULL(),
-			tCalendar.Job.EQ(mysql.String(job)),
-			tCalendar.SystemKind.EQ(
-				mysql.Int32(
-					int32(calendarresource.CalendarSystemKind_CALENDAR_SYSTEM_KIND_JOB_BIRTHDAYS),
-				),
-			),
-		),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	if existing != nil {
-		tx, err := s.db.BeginTx(ctx, nil)
-		if err != nil {
-			return nil, err
-		}
-		defer tx.Rollback()
-
-		tCalendar := table.FivenetCalendar
-		if _, err := tCalendar.
-			UPDATE(
-				tCalendar.Name,
-				tCalendar.Description,
-				tCalendar.Public,
-				tCalendar.Closed,
-				tCalendar.CreatorID,
-				tCalendar.CreatorJob,
-				tCalendar.SystemKind,
-			).
-			SET(
-				title,
-				mysql.String("System-managed birthday calendar"),
-				false,
-				true,
-				mysql.NULL,
-				mysql.String(job),
-				mysql.Int32(int32(calendarresource.CalendarSystemKind_CALENDAR_SYSTEM_KIND_JOB_BIRTHDAYS)),
-			).
-			WHERE(mysql.AND(
-				tCalendar.ID.EQ(mysql.Int64(existing.GetId())),
-			)).
-			LIMIT(1).
-			ExecContext(ctx, tx); err != nil {
-			return nil, err
-		}
-
-		if err := ensureBirthdayCalendarAccess(
-			ctx,
-			tx,
-			existing.GetId(),
-			job,
-			jobInfo,
-		); err != nil {
-			return nil, err
-		}
-
-		if err := tx.Commit(); err != nil {
-			return nil, err
-		}
-
-		return existing, nil
-	}
-
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return nil, err
-	}
-	defer tx.Rollback()
-
-	tCalendar := table.FivenetCalendar
-	stmt := tCalendar.
-		INSERT(
-			tCalendar.Job,
-			tCalendar.Name,
-			tCalendar.Description,
-			tCalendar.Public,
-			tCalendar.Closed,
-			tCalendar.Color,
-			tCalendar.CreatorID,
-			tCalendar.CreatorJob,
-			tCalendar.SystemKind,
-		).
-		VALUES(
-			mysql.String(job),
-			title,
-			mysql.String("System-managed birthday calendar"),
-			mysql.Bool(false),
-			mysql.Bool(true),
-			birthdayCalendarColor,
-			mysql.NULL,
-			mysql.String(job),
-			mysql.Int32(int32(calendarresource.CalendarSystemKind_CALENDAR_SYSTEM_KIND_JOB_BIRTHDAYS)),
-		).
-		ON_DUPLICATE_KEY_UPDATE(
-			tCalendar.Name.SET(mysql.String(title)),
-			tCalendar.Description.SET(mysql.String("System-managed birthday calendar")),
-			tCalendar.Public.SET(mysql.Bool(false)),
-			tCalendar.Closed.SET(mysql.Bool(true)),
-			tCalendar.DeletedAt.SET(mysql.TimestampExp(mysql.NULL)),
-			tCalendar.CreatorID.SET(mysql.IntExp(mysql.NULL)),
-			tCalendar.CreatorJob.SET(mysql.String(job)),
-			tCalendar.SystemKind.SET(mysql.Int32(int32(calendarresource.CalendarSystemKind_CALENDAR_SYSTEM_KIND_JOB_BIRTHDAYS))),
-		)
-
-	res, err := stmt.ExecContext(ctx, tx)
-	if err != nil {
-		return nil, err
-	}
-
-	lastID, err := res.LastInsertId()
-	if err != nil {
-		return nil, err
-	}
-
-	if lastID <= 0 {
-		tCalendar := table.FivenetCalendar.AS("calendar")
-		tCreator := table.FivenetUser.AS("creator")
-		tAvatar := table.FivenetFiles.AS("profile_picture")
-		row := &calendarresource.Calendar{}
-		selectStmt := tCalendar.
-			SELECT(
-				tCalendar.ID,
-				tCalendar.CreatedAt,
-				tCalendar.UpdatedAt,
-				tCalendar.DeletedAt,
-				tCalendar.Job,
-				tCalendar.Name,
-				tCalendar.Description,
-				tCalendar.Public,
-				tCalendar.Closed,
-				tCalendar.Color,
-				tCalendar.CreatorID,
-				tCreator.ID,
-				tCreator.Job,
-				tCreator.JobGrade,
-				tCreator.Firstname,
-				tCreator.Lastname,
-				tCreator.Dateofbirth,
-				tCreator.PhoneNumber,
-				tUserProps.AvatarFileID.AS("creator.profile_picture_file_id"),
-				tAvatar.FilePath.AS("creator.profile_picture"),
-				tCalendarSubs.CalendarID,
-				tCalendarSubs.UserID,
-				tCalendarSubs.CreatedAt,
-				tCalendarSubs.Confirmed,
-				tCalendarSubs.Muted,
-				tCalendar.SystemKind,
-			).
-			FROM(tCalendar.
-				LEFT_JOIN(tCreator,
-					tCalendar.CreatorID.EQ(tCreator.ID),
-				).
-				LEFT_JOIN(tUserProps,
-					tUserProps.UserID.EQ(tCalendar.CreatorID),
-				).
-				LEFT_JOIN(tCalendarSubs,
-					mysql.AND(
-						tCalendarSubs.CalendarID.EQ(tCalendar.ID),
-						tCalendarSubs.UserID.EQ(mysql.Int32(userInfo.GetUserId())),
-					),
-				).
-				LEFT_JOIN(tAvatar,
-					tAvatar.ID.EQ(tUserProps.AvatarFileID),
-				),
-			).
-			WHERE(mysql.AND(
-				tCalendar.DeletedAt.IS_NULL(),
-				tCalendar.Job.EQ(mysql.String(job)),
-				tCalendar.SystemKind.EQ(
-					mysql.Int32(
-						int32(
-							calendarresource.CalendarSystemKind_CALENDAR_SYSTEM_KIND_JOB_BIRTHDAYS,
-						),
-					),
-				),
-			)).
-			LIMIT(1)
-		if err := selectStmt.QueryContext(ctx, tx, row); err != nil {
-			if !errors.Is(err, qrm.ErrNoRows) {
-				return nil, err
-			}
-		}
-		if row.GetId() > 0 {
-			lastID = row.GetId()
-		}
-	}
-
-	if lastID <= 0 {
-		return nil, fmt.Errorf("unable to create birthday calendar for job %s", job)
-	}
-
-	if err := ensureBirthdayCalendarAccess(ctx, tx, lastID, job, jobInfo); err != nil {
-		return nil, err
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, err
-	}
-
-	return s.getCalendar(
-		ctx,
-		userInfo,
-		tCalendar.ID.EQ(mysql.Int64(lastID)),
-	)
-}
 
 func (s *Server) expandCalendarEntries(
 	ctx context.Context,
